move proto impl to its own module and add proptest
